+++ R/MAPAssignActivity.R
@@ -90,7 +90,8 @@ MAPAssignActivity <-
            seed                    = NULL,
            max.subsets             = 1000,
            use.sparse.assign       = FALSE,
-           drop.low.mut.samples    = TRUE) {
+           drop.low.mut.samples    = TRUE,
+           use.sig.presence.test   = FALSE) {
     if (drop.low.mut.samples) {
       spectra <- DropLowMutationSamples(spectra)
     } else {
@@ -115,7 +116,8 @@ MAPAssignActivity <-
         seed                    = seed,
         max.subsets             = max.subsets,
         use.sparse.assign       = use.sparse.assign,
-        drop.low.mut.samples    = drop.low.mut.samples)
+        drop.low.mut.samples    = drop.low.mut.samples, 
+        use.sig.presence.test   = use.sig.presence.test)
 
       return(retval1)
     }