+++ R/RunMAPOneSample.R
@@ -24,7 +24,8 @@ RunMAPOnOneSample <-
            seed                    = NULL,
            max.subsets             = 1000,
            use.sparse.assign       = FALSE,
-           drop.low.mut.samples    = TRUE) {
+           drop.low.mut.samples    = TRUE,
+           use.sig.presence.test   = FALSE) {
     
     if (drop.low.mut.samples) {
       spect <- DropLowMutationSamples(spect)
@@ -49,7 +50,8 @@ RunMAPOnOneSample <-
       seed                    = seed,
       max.subsets             = max.subsets,
       use.sparse.assign       = use.sparse.assign,
-      drop.low.mut.samples    = drop.low.mut.samples)
+      drop.low.mut.samples    = drop.low.mut.samples, 
+      use.sig.presence.test   = use.sig.presence.test)
     
     if (!is.null(retval$error.messages)) {
       return(retval)